refactor: Rename `categoryFilter.tsx` for consistency.
